move interceptor impl out of lib.rs
